Fixed a bug in the CanOpenNodeExporter_V4.Make_cname.
The method would fail on a array out-of-bounds exception. This was do to and empty
string that was added to the tokens list. It was fixed by removing all empty strings
from the tokens list as was done in CanOpenNodeExporter.make_cname.

The conditions in the test for this method all failed. The method was modified
to pass the conditions in the test.